% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_barplot.R
\name{create_barplot}
\alias{create_barplot}
\title{Create and save a customized barplot for GSEA results}
\usage{
create_barplot(
  data,
  output_path,
  custom_labels,
  height = 49,
  width = 30,
  axis_y = "NES"
)
}
\arguments{
\item{data}{A data frame containing GSEA results with columns such as \code{datatype}, \code{NES}, \code{-Log10FDR}, and \code{New_name}.}

\item{output_path}{The file path where the barplot will be saved (SVG format).}

\item{custom_labels}{A named vector of custom expressions for x-axis labels.}

\item{height}{Height of the saved plot in inches. Default: 49.}

\item{width}{Width of the saved plot in inches. Default: 30.}

\item{axis_y}{Name of the column to use for the y-axis aesthetic, as a string. Default: "NES".}
}
\description{
Create and save a customized barplot for GSEA results
}
\details{
This function generates a customized barplot with grouped bars,
adjusted aesthetics, and personalized axis labels, saving the result as an SVG file.
}
