% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_barplot.R
\name{gsea_barplot}
\alias{gsea_barplot}
\title{Create and save a customized barplot for GSEA results}
\usage{
gsea_barplot(
  data,
  output_path,
  custom_labels,
  height = 49,
  width = 30,
  axis_y = "NES"
)
}
\arguments{
\item{data}{A data frame containing GSEA results with columns such as \code{datatype}, \code{NES}, \code{-Log10FDR}, and \code{New_name}.}

\item{output_path}{The file path where the barplot will be saved (SVG format).}

\item{custom_labels}{A named vector of custom expressions for x-axis labels.}

\item{height}{Height of the saved plot in inches. Default: 49.}

\item{width}{Width of the saved plot in inches. Default: 30.}

\item{axis_y}{Name of the column to use for the y-axis aesthetic, as a string. Default: "NES".}
}
\description{
This function generates a customized barplot with:
\itemize{
\item Grouped bars.
\item Adjusted aesthetics.
\item Personalized axis labels.
\item Optionally save the result in SVG format.
}
}
