% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_circos.R
\name{is_circos_genome}
\alias{is_circos_genome}
\title{Test whether an object is a circos genome descriptor}
\usage{
is_circos_genome(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Logical scalar. \code{TRUE} if \code{x} inherits from class
\code{"circos_genome"}, otherwise \code{FALSE}.
}
\description{
Test whether an object is a circos genome descriptor
}
\examples{
ce10 <- resolve_genome("ce10")
is_circos_genome(ce10)
is_circos_genome(data.frame())

}
\seealso{
\code{\link[=resolve_genome]{resolve_genome()}}, \code{\link[=nice_circos]{nice_circos()}}
}
