% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsKit-package.R
\docType{package}
\name{OmicsKit-package}
\alias{OmicsKit}
\alias{OmicsKit-package}
\title{OmicsKit: A bioinformatics toolset for multiomics analysis}
\description{
This package contains functions that help in manipulating tables and generating plots for multi-omics analysis including genomics, transcriptomics, proteomics, methylomics and immunoinformatics.
}
\author{
\strong{Maintainer}: David Requena \email{david.requena@nyulangone.org} (\href{https://orcid.org/0000-0002-5968-1133}{ORCID})

Authors:
\itemize{
  \item Daniel Guevara \email{dgdiaz011202@gmail.com} (\href{https://orcid.org/0009-0001-2786-8729}{ORCID})
}

}
\keyword{internal}
