% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_circos.R
\name{print.circos_genome}
\alias{print.circos_genome}
\title{Print a circos genome descriptor}
\usage{
\method{print}{circos_genome}(x, ...)
}
\arguments{
\item{x}{A \code{"circos_genome"} object returned by \code{\link[=resolve_genome]{resolve_genome()}}.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a circos genome descriptor
}
\examples{
ce10 <- resolve_genome("ce10")
print(ce10)

}
\seealso{
\code{\link[=resolve_genome]{resolve_genome()}}
}
