% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stars.R
\name{get_stars}
\alias{get_stars}
\title{Obtain annotations to display significance.}
\usage{
get_stars(geneID, object, thresholds = c(0.001, 0.01, 0.1, 0.25))
}
\arguments{
\item{geneID}{Ensembl ID of the gene of interest.}

\item{object}{DESeq2 results object of a comparison.}

\item{thresholds}{Vector with 4 values of significance. Default c(0.001, 0.01, 0.1, 0.25).}
}
\value{
A single character string: \code{"****"}, \code{"***"}, \code{"**"}, \code{"*"},
\code{"ns"} (not significant), or \code{"Gene ID not found"} if the gene is absent
from \code{object}.
}
\description{
This function will create asteriscs (*) from DESeq2 results objects to represent the significance of comparisons.
}
\examples{
data(deseq2_results)

# get_stars expects a column named "ensembl"
res <- deseq2_results
colnames(res)[colnames(res) == "gene_id"] <- "ensembl"

# Get significance stars for the most significant gene
get_stars(
  geneID = res$ensembl[1],
  object = res
)

# Custom thresholds
get_stars(
  geneID     = res$ensembl[1],
  object     = res,
  thresholds = c(0.001, 0.01, 0.05, 0.10)
)

# Non-significant gene
get_stars(
  geneID = res$ensembl[nrow(res)],
  object = res
)

}
\seealso{
\code{\link[=detect_filter]{detect_filter()}} to identify detectable genes before annotating;
\code{\link[=nice_VSB]{nice_VSB()}} where significance stars can be added to plots;
\link{deseq2_results} for an example input.
}
