% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_PA.R
\name{merge_PA}
\alias{merge_PA}
\title{Merge GSEA results data frames.}
\usage{
merge_PA(input_directory, output_file = "collections_merged_gsea_data.tsv")
}
\arguments{
\item{input_directory}{The directory containing the GSEA collection results in TSV format.}

\item{output_file}{The output file to save the merged data. If not provided, the file will be saved in the input directory.}
}
\description{
After running GSEA_all.sh from GSEA.sh, merge_PA function joins .tsv files to a single file
}
