% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsKit-package.R
\docType{package}
\name{OmicsKit-package}
\alias{OmicsKit}
\alias{OmicsKit-package}
\title{OmicsKit: A bioinformatics toolset for multiomics analysis}
\description{
This package contains functions that help in manipulating tables and generating plots for multi-omics analysis including genomics, transcriptomics, proteomics, methylomics and immunoinformatics.
}
\author{
\strong{Maintainer}: Daniel Guevara \email{dgdiaz011202@gmail.com} (\href{https://orcid.org/0009-0001-2786-8729}{ORCID})

Authors:
\itemize{
  \item Daniel Garbozo \email{dan.garbozo.1201.urp@gmail.com} (\href{https://orcid.org/0009-0003-2495-6568}{ORCID})
  \item Angela Alarcon \email{alarcondcg@gmail.com} (\href{https://orcid.org/0000-0003-0293-5603}{ORCID})
  \item Jennifer Almeyda \email{jalmeydatz@gmail.com} (\href{https://orcid.org/0009-0003-9589-1167}{ORCID})
  \item David Requena \email{david.requena@nyulangone.org} (\href{https://orcid.org/0000-0002-5968-1133}{ORCID})
}

}
\keyword{internal}
