% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GSEA.R
\name{plot_GSEA}
\alias{plot_GSEA}
\title{Plot global GSEA results}
\usage{
plot_GSEA(
  data,
  geneset_col,
  collection_col,
  nes_col,
  logfdr_col,
  text_size_genesets = 5,
  text_size_collection = 5
)
}
\arguments{
\item{data}{Data frame containing the GSEA results.}

\item{geneset_col}{Name of the column containing the genesets.}

\item{collection_col}{Name of the column containing the collections.}

\item{nes_col}{Name of the column containing the NES values.}

\item{logfdr_col}{Name of the column containing \eqn{-\log_{10}(FDR)} values.}

\item{text_size_genesets}{Text size for the geneset labels.}

\item{text_size_collection}{Text size for the collection labels.}
}
\value{
GSEA barplots arranged in a grid.
}
\description{
Generates a composite plot displaying NES values, pathway labels,
and a \emph{logFDR} legend, organized by MSigDB collections.
}
